+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2023-07-06
+
+### Added
+
+- Selected item font size and weight
+
+### Fixed
+
+- Selected item color is in use
+
 ## [1.2.0] - 2023-07-02
 
 ### Added
@@ -62,7 +72,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use regular expression for URL matching
 - Configure popup autoclosing and autoclosing time
 
-[Unreleased]: https://github.com/vikian050194/click/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/vikian050194/click/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/vikian050194/click/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/vikian050194/click/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/vikian050194/click/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/vikian050194/click/compare/v0.4.0...v1.0.0